requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
openpyxl==3.1.5
xhtml2pdf==0.2.14
